feat: 🎸 Added `types` package
