+++ packages/types/tsconfig-build.json
@@ -1,0 +1,4 @@
+{
+  "extends": "./tsconfig.json",
+  "exclude": ["./vite.config.ts"]
+}
